moved config plugins to separate folderand added purging to plugin css files
